ocr_service/requirements.txt
flask
pillow
pytesseract
sqlalchemy
pymysql
mysqlclient